generator/requirements.txt
websockets==12.0
azure-eventhub==5.11.6
aiohttp==3.9.5
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
aiokafka==0.10.0
structlog==24.2.0
